+++ .github/workflows/test-mango.yml
@@ -14,12 +14,11 @@ permissions:
 
 jobs:
   build:
-    runs-on: ${{ matrix.os }}
+    runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
         python-version: ["3.10", "3.11", "3.12", "3.13"]
-        os: [windows-latest, macOS-latest, ubuntu-latest]
     steps:
     - uses: actions/checkout@v4
     - name: Set up Python
