+++ mango/messages/codecs.py
@@ -14,8 +14,6 @@ https://gitlab.com/sscherfke/aiomas/
 import inspect
 import json
 
-import msgspec
-
 from mango.messages.message import (
     ACLMessage,
     MangoMessage,
@@ -179,25 +177,6 @@ class JSON(Codec):
 
     def decode(self, data):
         return json.loads(data.decode(), object_hook=self.deserialize_obj)
-
-
-class FastJSON(Codec):
-    def __init__(self):
-        super().__init__()
-        self.add_serializer(*ACLMessage.__json_serializer__())
-        self.add_serializer(*MangoMessage.__json_serializer__())
-        self.add_serializer(*enum_serializer(Performatives))
-
-        self.encoder = msgspec.json.Encoder(enc_hook=self.serialize_obj)
-        self.decoder = msgspec.json.Decoder(
-            dec_hook=lambda _, b: self.deserialize_obj(b), type=MangoMessage
-        )
-
-    def encode(self, data):
-        return self.encoder.encode(data)
-
-    def decode(self, data):
-        return self.decoder.decode(data)
 
 
 class PROTOBUF(Codec):
